more functionality into handler
